move sfh functions to new file
